refactor: Remove diner service abstract
- For more simple code
- Abstract was only for calling check method of diner, company and member.
